refactor: Filter 컴포넌트 별도 컴포넌트로 분리 (#55)
